+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## master
 
 * Added History panel to iceleaf GUI
+* Added SubscribeAddressUpdates gRPC api to node to monitor addresses
 
 ## 1.7.0